+++ pkg/manifests/config.go
@@ -38,6 +38,26 @@ type Config struct {
 	UserWorkloadConfiguration      *UserWorkloadConfiguration      `json:"-"`
 }
 
+// GetPrometheusUWAdditionalAlertmanagerConfigs returns the alertmanager configurations for
+// the User Workload Monitoring Prometheus instance.
+// If no additional configurations are specified, GetPrometheusUWAdditionalAlertmanagerConfigs returns an empty array.
+func (c Config) GetPrometheusUWAdditionalAlertmanagerConfigs() []AdditionalAlertmanagerConfig {
+	if c.UserWorkloadConfiguration == nil {
+		return []AdditionalAlertmanagerConfig{}
+	}
+
+	if c.UserWorkloadConfiguration.Prometheus == nil {
+		return []AdditionalAlertmanagerConfig{}
+	}
+
+	alertmanagerConfigs := c.UserWorkloadConfiguration.Prometheus.AlertmanagerConfigs
+	if alertmanagerConfigs == nil {
+		return []AdditionalAlertmanagerConfig{}
+	}
+
+	return alertmanagerConfigs
+}
+
 type ClusterMonitoringConfiguration struct {
 	PrometheusOperatorConfig *PrometheusOperatorConfig    `json:"prometheusOperator"`
 	PrometheusK8sConfig      *PrometheusK8sConfig         `json:"prometheusK8s"`
@@ -403,6 +423,7 @@ type PrometheusRestrictedConfig struct {
 	VolumeClaimTemplate *monv1.EmbeddedPersistentVolumeClaim `json:"volumeClaimTemplate"`
 	RemoteWrite         []monv1.RemoteWriteSpec              `json:"remoteWrite"`
 	EnforcedSampleLimit *uint64                              `json:"enforcedSampleLimit"`
+	AlertmanagerConfigs []AdditionalAlertmanagerConfig       `json:"additionalAlertManagerConfigs"`
 }
 
 func (u *UserWorkloadConfiguration) applyDefaults() {